ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
onnxruntime==1.16.3
opencv-python==4.8.1.78
pytest==7.4.3
pytest-asyncio==0.21.1